Changed type of set_intensity() arg back.

+++ components/led_display/led_display.h
@@ -77,7 +77,7 @@ public:
 
   void loop() override;
 
-  void set_intensity(uint intensity) {
+  void set_intensity(uint8_t intensity) {
     this->intensity_ = intensity;
     this->brightness_ = map(intensity, 0, 100, MIN_LEDS_ON_DELAY, MAX_LEDS_ON_DELAY);
     ESP_LOGV(TAG, "Intensity: %u\nBrightness: %u", this->intensity_, this->brightness_);
